+++ hello/src/AppBundle/Repository/UsergroupRepository.php
@@ -23,13 +23,41 @@ class UsergroupRepository
 	public function deleteUsergroup($id)
 	{
 		$usergroup = $this->usergroupRepo->find($id);
+		if($usergroup==null)
+		{
+			return "There is no user group with id ".$id.".";
+		}
+
 		$this->em->remove($usergroup);
 		$this->em->flush();
 	}
 
 	public function getUsergroupById($id)
 	{
-		return $this->usergroupRepo->find($id);
+		try{
+			$qb = $this->usergroupRepo->createQueryBuilder("u");
+			$qb->where($qb->expr()->eq("u.userGroupId",":userGroupId"));
+
+			$qb->setParameters(array(
+				"userGroupId"=> $id));
+			
+			return $qb->getQuery()->getSingleResult();
+		}
+		catch(\Exception $e)
+		{
+			return null;
+		}
+	}
+
+	public function getUsergroupByIdOrDefault($id)
+	{
+		$usergroup = $this->getUsergroupById($id);
+		if($usergroup == null)
+		{
+			$usergroup = $this->getDefaultUsergroup();
+		}
+
+		return $usergroup;
 	}
 
 	public function editUsergroup($usergroup, $modifiedUsergroup)
@@ -43,6 +71,13 @@ class UsergroupRepository
 	public function getAllUsergroup()
 	{
 		return $this->usergroupRepo->findAll();
+	}
+
+	public function getDefaultUsergroup()
+	{
+		$qb = $this->usergroupRepo->createQueryBuilder("u")->setMaxResults(1);
+		$result = $qb->getQuery()->getResult();
+		return $result;
 	}
 
 }
